anhnt update phân quyền chia quyền admin và staff

+++ routes/web.php
@@ -45,7 +45,7 @@ Route::prefix('admin')->name('admin.')->group(function () {
 
 
 // Routes for authenticated users with 'admin' role
-Route::prefix('admin')->name('admin.')->middleware('auth', 'role:admin')->group(function () {
+Route::prefix('admin')->name('admin.')->middleware('auth', 'role:admin,staff')->group(function () {
     Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
 
     // CRUD routes for User management
@@ -56,7 +56,7 @@ Route::prefix('admin')->name('admin.')->middleware('auth', 'role:admin')->group(
     Route::get('/taikhoans/{id}', [UserController::class, 'show'])->name('taikhoans.show'); // Show user details
     Route::get('/taikhoans/{id}/edit', [UserController::class, 'edit'])->name('taikhoans.edit'); // Edit user
     Route::put('/taikhoans/{id}', [UserController::class, 'update'])->name('taikhoans.update'); // Update user
-    Route::delete('/taikhoans/{id}', [UserController::class, 'destroy'])->name('taikhoans.destroy'); // Delete user
+    Route::delete('/taikhoans/{id}', [UserController::class, 'destroy'])->middleware('auth', 'role:admin')->name('taikhoans.destroy'); // Delete user
     Route::get('/profile', [UserController::class, 'profile'])->name('profile');
     Route::put('/update/{id}', [UserController::class, 'updateProfile'])->name('updateProfile');
     Route::put('/profile/updatePassword', [UserController::class, 'updatePassword'])->name('profile.updatePassword');
@@ -118,13 +118,13 @@ Route::prefix('admin')->name('admin.')->middleware('auth', 'role:admin')->group(
     // sản phẩm
     Route::prefix('sanphams')->name('sanphams.')->group(function () {
         Route::get('/', [SanPhamController::class, 'index'])->name('index');
-        Route::get('create', [SanPhamController::class, 'create'])->name('create');
-        Route::post('store', [SanPhamController::class, 'store'])->name('store');
+        Route::get('create', [SanPhamController::class, 'create'])->middleware('auth', 'role:admin')->name('create');
+        Route::post('store', [SanPhamController::class, 'store'])->middleware('auth', 'role:admin')->name('store');
         Route::get('/{id}/show', [SanPhamController::class, 'show'])->name('show');
-        Route::get('/{id}/edit', [SanPhamController::class, 'edit'])->name('edit');
+        Route::get('/{id}/edit', [SanPhamController::class, 'edit'])->middleware('auth', 'role:admin')->name('edit');
         Route::put('/{id}/update', [SanPhamController::class, 'update'])->name('update');
-        Route::delete('/{id}/destroy', [SanPhamController::class, 'destroy'])->name('destroy');
-        Route::post('/{id}/restore', [SanPhamController::class, 'restore'])->name('restore');
+        Route::delete('/{id}/destroy', [SanPhamController::class, 'destroy'])->middleware('auth', 'role:admin')->name('destroy');
+        Route::post('/{id}/restore', [SanPhamController::class, 'restore'])->middleware('auth', 'role:admin')->name('restore');
         Route::get('/sanpham/{id}/filterDanhGia/{star}', [SanPhamController::class, 'filterDanhGia'])->name('filterDanhGia');
     });
 
@@ -163,13 +163,13 @@ Route::prefix('admin')->name('admin.')->middleware('auth', 'role:admin')->group(
 });
 
 // Routes for authenticated users with 'staff' role
-Route::prefix('staff')->name('staff.')->middleware('auth', 'role:staff')->group(function () {
-    Route::get('/dashboard', [StaffDashboardController::class, 'index'])->name('dashboard');
-    Route::get('/khachhangs', [UserController::class, 'khachhangs'])->name('khachhangs'); // Display users list
-    Route::get('/khachhang/{id}', [UserController::class, 'show'])->name('taikhoans.show'); // Show user details
+// Route::prefix('staff')->name('staff.')->middleware('auth', 'role:staff')->group(function () {
+//     Route::get('/dashboard', [StaffDashboardController::class, 'index'])->name('dashboard');
+//     Route::get('/khachhangs', [UserController::class, 'khachhangs'])->name('khachhangs'); // Display users list
+//     Route::get('/khachhang/{id}', [UserController::class, 'show'])->name('taikhoans.show'); // Show user details
+// });
 
 
-});
 // Routes for unauthenticated users
 Route::prefix('customer')->name('customer.')->group(function () {
     Route::get('login', [CustomerLoginController::class, 'showLoginForm'])->name('login');
